Input Manager
Build an input manager for mobile. it only takes 2 inputs for now and
they are both holds. The screen is split in an X and it works right now.
Code needs a good clean up and some optimization.  I'll eventually work
on a Loggin system and some debug tools for mobile. I'll explore more
options for inputs later on.