Dialogue Portraits and Player Name
Added support for the portraits and for the player name. The Dialogue
Library takes care of everything and parses the text. The Dialogue
Manager recieves at set of lines with attached the enum of the speaker.

Two main issues remain:   Dialogue manager methods for getting the next
line and changing the game state to Dialogue/Paused.