Unity Editor Handles
Worked on the editor Script. Mostly on getting the unity Handles to draw
and making sure you cannot edit the mesh through the transform
component. (probably going to add support for that later on). Need to
get the handles to affect the mesh and then the implementation of the
triangulation algorithm.